rename files to match other projects
