+++ src/lib/public_api.ts
@@ -2,10 +2,10 @@ export * from './portal/portal';
 export * from './overlay/overlay';
 export * from './overlay/overlay-container';
 export * from './overlay/overlay-ref';
-export * from './toastr/toast-directive';
-export * from './toastr/toast-component';
-export * from './toastr/toastr-service';
+export * from './toastr/toast.directive';
+export * from './toastr/toast.component';
+export * from './toastr/toastr.service';
 export * from './toastr/toastr-config';
-export * from './toastr/toastr-module';
+export * from './toastr/toastr.module';
 export * from './toastr/toast-injector';
 export * from './toastr/toast-token';
