+++ src/lib/toastr/toast.component.ts
